New Program start in docs dir and valid name check
New Program button opens save as dialog box in user's Documents/Code12 Programs/ directory, or creates it if it doesn't exist.
If new program path is invalid (doesn't end in .java, is a known class name, doesn't start with uppercase) an error message is displayed and the Save As dialog is reopened.